[runtime] Add first cache: GetNamedPropertyCache (#425)
## Summary

Implement the first property cache, for the `GetNamedProperty`
instruction. This uses the recently implemented shapes and prototype
chain validity guards to cache the location of property lookups. Own
properties only need to cache the shape while prototype properties need
to cache the prototype object, shape, and validity guard.

We have a fast path now to check the `GetNamedPropertyCache` for the
instruction (lazily initialized). On a cache miss we perform the
existing full property lookup, then write the cached location if
desired.

Set up general infra for adding more caches:
- Cache indices and a cache array generated for bytecode function
- Wrote simple `ValidityGuard` wrapper
- `Cache` set up to be extended with new variants as caches are added

Polymorphic caches are currently not supported (nor are megamorphic
caches of any form). The first time we see multiple shapes we stop
caching entirely.

Significant performance impact. This instruction cache alone is +11% on
octane.

| Benchmark | Before | After | Change |
|:---|---:|---:|---:|
| Richards | 399 | 557 | +39.6% |
| DeltaBlue | 572 | 804 | +40.6% |
| Crypto | 490 | 531 | +8.4% |
| RayTrace | 1,106 | 1,371 | +24.0% |
| EarleyBoyer | 1,707 | 1,888 | +10.6% |
| RegExp | 356 | 367 | +3.1% |
| Splay | 2,369 | 2,318 | -2.2% |
| SplayLatency | 5,952 | 6,132 | +3.0% |
| NavierStokes | 608 | 627 | +3.1% |
| PdfJS | 2,489 | 2,674 | +7.4% |
| Mandreel | 257 | 264 | +2.7% |
| MandreelLatency | 1,623 | 1,678 | +3.4% |
| Gameboy | 2,337 | 2,384 | +2.0% |
| CodeLoad | 38,355 | 38,155 | -0.5% |
| Box2D | 1,757 | 2,588 | +47.3% |
| zlib | 843 | 842 | -0.1% |
| Typescript | 7,923 | 9,369 | +18.3% |
| **Total (octane)** | **1,470** | **1,638** | **+11.4%** |

## Tests

- CI
- Assorted integration tests that hit various `GetNamedProperty` cache
paths
- Updated all bytecode snapshot instructions that used
`GetNamedProperty` to include the cache index

+++ src/js/runtime/transitions.rs
@@ -21,7 +21,7 @@ impl PropertyDefinition {
 /// Where a property is stored on an object.
 pub enum PropertyLocation {
     /// Property is stored in the named properties array at this index.
-    PropertyArray { index: u32 },
+    PropertyArray { index: u16 },
 }
 
 #[derive(Clone, Copy)]
